Refactor RequestContextPanel formatters
